+++ tests/test_cases/test_cocotb/test_scheduler.py
@@ -470,7 +470,7 @@ async def test_test_repr(_):
     """Test RunningTest.__repr__"""
     log = logging.getLogger("cocotb.test")
 
-    current_test = cocotb.scheduler._test
+    current_test = cocotb._scheduler._test
     log.info(repr(current_test))
     assert re.match(
         r"<Test test_test_repr running coro=test_test_repr\(\)>", repr(current_test)
@@ -488,7 +488,7 @@ class TestClassRepr(Coroutine):
     async def check_repr(self, dut):
         log = logging.getLogger("cocotb.test")
 
-        current_test = cocotb.scheduler._test
+        current_test = cocotb._scheduler._test
         log.info(repr(current_test))
         assert re.match(
             r"<Test TestClassRepr running coro=TestClassRepr\(\)>", repr(current_test)
@@ -535,7 +535,7 @@ async def test_start_soon_scheduling(dut):
         log = logging.getLogger("cocotb.test")
         log.debug("react_wrapper start")
         assert coro_scheduled is False
-        cocotb.scheduler._react(trigger)
+        cocotb._scheduler._react(trigger)
         assert coro_scheduled is True
         log.debug("react_wrapper end")
 
@@ -665,7 +665,7 @@ async def test_start_scheduling(dut):
         log = logging.getLogger("cocotb.test")
         log.debug("react_wrapper start")
         sim_resumed = False
-        cocotb.scheduler._react(trigger)
+        cocotb._scheduler._react(trigger)
         sim_resumed = True
         log.debug("react_wrapper end")